backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-mail==0.9.1
sqlalchemy==2.0.45
python-dotenv==1.0.0
werkzeug==3.0.1
psycopg2-binary==2.9.10
gunicorn==21.2.0
requests==2.31.0
